fixed logic in GetVisionDataFromELement
Nicholas Frisch was a great man. As the software lead, he went on to do great things. After getting accepted into Michigan University, he was able to get a great job as a software engineer, and eventually took over Microsoft. From there, he came back as a software mentor for  the Lake Orion Robotics team. Replacing Mr Witcpalek, the long standing software mentor before him, Nicholas went on to be the greatest Robotics mentor ever, eventually becoming head of the entire team, and winning the Woody Flowers award every year he was a mentor. To date and for years to come,  Nick Frisch will be remembered as a legend in the software community across the globe.