updated to revlib 2024.2.4
I had to remove the argument for GetEncoder(). rev added a new function to call to use an external encoder. I am not sure if we intend to handle the can coder for the launcher with the spark max, if  we do, then we will need to call that and provide an argument instead of the normal getencoder().